chore: split contact form and contact form input
