+++ src/components/ContactCard/ContactForm.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Form, Field } from "react-final-form";
-import Icon from "cozy-ui/react/Icon";
+import { Form } from "react-final-form";
+import ContactFieldForm from "./ContactFieldForm";
 import { Button } from "cozy-ui/react/Button";
 import { translate } from "cozy-ui/react/I18n";
 
@@ -64,85 +64,6 @@ const fields = [
     type: "textarea"
   }
 ];
-
-const getInputComponent = inputType =>
-  inputType === "textarea" ? "textarea" : "input";
-
-class ContactFieldInput extends React.Component {
-  state = {
-    renderLabel: false
-  };
-
-  showLabel = () => {
-    this.setState({
-      renderLabel: true
-    });
-  };
-
-  hideLabelIfEmpty = e => {
-    this.setState({
-      renderLabel: e.target.value
-    });
-  };
-
-  render() {
-    const { name, type, withLabel } = this.props;
-    const { renderLabel } = this.state;
-
-    return (
-      <div className="contact-form__input-wrapper">
-        <Field
-          name={name}
-          type={type}
-          onFocus={this.showLabel}
-          onBlur={this.hideLabelIfEmpty}
-          component={getInputComponent(type)}
-          className="contact-form__input"
-        />
-        {withLabel &&
-          renderLabel && (
-            <Field
-              name={`${name}Label`}
-              type="text"
-              component="input"
-              className="contact-form__label-input"
-            />
-          )}
-      </div>
-    );
-  }
-}
-ContactFieldInput.propTypes = {
-  name: PropTypes.string.isRequired,
-  type: PropTypes.string.isRequired,
-  withLabel: PropTypes.bool
-};
-ContactFieldInput.defaultProps = {
-  hasLabel: false
-};
-
-const ContactFieldForm = ({ icon, name, type, label, inputWithLabel }) => (
-  <div className="contact-form__field">
-    <label className="contact-form__label">
-      {icon && (
-        <Icon icon={icon} color="coolGrey" className="contact-form__icon" />
-      )}
-      {label}
-    </label>
-    <ContactFieldInput name={name} type={type} withLabel={inputWithLabel} />
-  </div>
-);
-ContactFieldForm.propTypes = {
-  icon: PropTypes.any, // shall be a SVG prop type
-  name: PropTypes.string.isRequired,
-  type: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  inputWithLabel: PropTypes.bool
-};
-ContactFieldForm.defaultProps = {
-  icon: null,
-  hasLabel: false
-};
 
 class ContactForm extends React.Component {
   formDataToContact = data => {
@@ -219,9 +140,9 @@ class ContactForm extends React.Component {
           <div>
             <form onSubmit={handleSubmit} className="contact-form">
               <div className="contact-form__fields">
-                {fields.map(({ name, icon, type, hasLabel }, index) => (
+                {fields.map(({ name, icon, type, hasLabel }) => (
                   <ContactFieldForm
-                    key={index}
+                    key={name}
                     icon={icon}
                     name={name}
                     label={t(`field.${name}`)}
